refactor: remove deprecated settings module
