+++ ingenious/db/azuresql/__init__.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, Optional
 
 import pyodbc
 
-from ingenious.config.settings import IngeniousSettings
+from ingenious.config import IngeniousSettings
 
 # Future import placeholders for advanced error handling
 # from ingenious.core.error_handling import (
